Implement the card in RECORD SPEED
