+++ app/src/main/res/layout/contract_detail_activity.xml
@@ -1,7 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical">
@@ -32,8 +31,8 @@
             <com.google.android.material.appbar.CollapsingToolbarLayout
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
-                android:minHeight="@dimen/tab_layout_height"
                 android:background="@color/colorSurface"
+                android:minHeight="@dimen/tab_layout_height"
                 app:layout_scrollFlags="scroll|exitUntilCollapsed|snap">
 
                 <LinearLayout
@@ -41,94 +40,9 @@
                     android:layout_height="wrap_content"
                     android:orientation="vertical">
 
-                    <com.google.android.material.card.MaterialCardView
-                        android:layout_marginStart="@dimen/base_margin_double"
-                        android:layout_marginEnd="@dimen/base_margin_double"
-                        android:layout_marginTop="@dimen/base_margin"
-                        android:layout_marginBottom="@dimen/base_margin_triple"
-                        android:layout_height="@dimen/contract_card_height"
-                        app:strokeWidth="1dp"
-                        app:strokeColor="@color/color_divider"
-                        android:layout_width="match_parent"
-                        android:theme="@style/ThemeOverlay.Hedvig.InsuranceCard">
-
-                        <androidx.constraintlayout.widget.ConstraintLayout
-                            android:id="@+id/container"
-                            android:transitionName="contract_card"
-                            android:layout_width="match_parent"
-                            android:layout_height="match_parent"
-                            android:background="@color/hedvig_light_gray"
-                            android:padding="@dimen/base_margin_double">
-
-                            <TextView
-                                android:id="@+id/firstStatusPill"
-                                android:layout_width="wrap_content"
-                                android:layout_height="wrap_content"
-                                android:background="@drawable/background_pill"
-                                android:backgroundTint="@color/yellow"
-                                android:visibility="gone"
-                                app:layout_constraintTop_toTopOf="parent"
-                                app:layout_constraintStart_toStartOf="parent"
-                                tools:text="AWAITING START DATE"
-                                tools:visibility="visible" />
-
-                            <TextView
-                                android:id="@+id/secondStatusPill"
-                                android:layout_width="wrap_content"
-                                android:layout_height="wrap_content"
-                                android:background="@drawable/background_pill"
-                                android:backgroundTint="@color/yellow"
-                                android:visibility="gone"
-                                android:layout_marginTop="@dimen/base_margin"
-                                app:layout_constraintTop_toBottomOf="@id/firstStatusPill"
-                                app:layout_constraintStart_toStartOf="parent"
-                                tools:text="AWAITING START DATE"
-                                tools:visibility="visible" />
-
-                            <ImageView
-                                android:layout_width="@dimen/contract_card_logo_size"
-                                android:layout_height="@dimen/contract_card_logo_size"
-                                app:layout_constraintTop_toTopOf="parent"
-                                app:layout_constraintEnd_toEndOf="parent"
-                                app:srcCompat="@drawable/ic_hedvig_h"
-                                app:tint="@color/lavender_900"
-                                android:contentDescription="@null" />
-
-                            <TextView
-                                android:id="@+id/contractName"
-                                android:textAppearance="?textAppearanceHeadline6"
-                                android:layout_width="wrap_content"
-                                android:layout_height="wrap_content"
-                                app:layout_constraintBottom_toTopOf="@id/contractPills"
-                                android:layout_marginBottom="@dimen/base_margin"
-                                app:layout_constraintStart_toStartOf="parent"
-                                tools:text="Home Insurance" />
-
-                            <androidx.recyclerview.widget.RecyclerView
-                                android:id="@+id/contractPills"
-                                android:layout_width="0dp"
-                                android:layout_height="wrap_content"
-                                app:layout_constraintBottom_toBottomOf="parent"
-                                app:layout_constraintStart_toStartOf="parent"
-                                app:layoutManager="com.google.android.flexbox.FlexboxLayoutManager"
-                                app:layout_constraintEnd_toStartOf="@+id/arrow"
-                                android:orientation="horizontal"
-                                android:layout_marginEnd="@dimen/base_margin"
-                                tools:itemCount="6"
-                                tools:listitem="@layout/contract_pill" />
-
-
-                            <ImageView
-                                android:layout_width="wrap_content"
-                                android:layout_height="wrap_content"
-                                app:srcCompat="@drawable/ic_arrow_forward"
-                                app:layout_constraintEnd_toEndOf="parent"
-                                app:layout_constraintBottom_toBottomOf="parent"
-                                android:id="@+id/arrow" />
-
-                        </androidx.constraintlayout.widget.ConstraintLayout>
-
-                    </com.google.android.material.card.MaterialCardView>
+                    <include
+                        android:id="@+id/card"
+                        layout="@layout/insurance_contract_card" />
 
                     <com.google.android.material.tabs.TabLayout
                         android:id="@+id/tabContainer"
